requirements.txt
Flask>=2.0,<3.0
python-dotenv>=1.0,<2.0
gspread>=5.0,<6.0
oauth2client>=4.1,<5.0
gunicorn==21.2.0
sqlalchemy>=1.4,<2.0
Flask-SQLAlchemy>=2.5,<3.0